split into components
