+++ .pre-commit-config.yaml
@@ -1,3 +1,6 @@
+ci:
+  autoupdate_schedule: monthly
+
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.4.0
@@ -52,7 +55,7 @@ repos:
     hooks:
       - id: mypy
         additional_dependencies: [
-          "git+https://github.com/python-cosmology/cosmology-api.git#egg=cosmology-api"
+          "git+https://github.com/cosmology-api/cosmology-api.git#egg=cosmology-api"
         ]
         args:
           - --strict
