+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-  name = "cosmology.core"
+  name = "cosmology"
   version = "0.1.0"
   description = "Cosmology API standard compatability library"
   readme = "README.rst"
@@ -22,7 +22,7 @@
     "Programming Language :: Python :: 3",
   ]
   dependencies = [
-    "cosmology.api @ git+https://github.com/python-cosmology/cosmology-api.git",
+    "cosmology.api @ git+https://github.com/cosmology-api/cosmology-api.git",
   ]
 
 [project.optional-dependencies]
